feat: 调整auth的url，react部署，nginx脚本提交
