+++ doc/deploy/nginx/conf/nginx.conf
@@ -51,26 +51,6 @@ http  {
         gzip_vary                   on;
         gzip_types                  text/plain text/css application/json application/x-javascript text/xml application/xml application/xml+rss text/javascript;
 
-        location  /swagger-ui.html  {
-            proxy_pass              https://api.laokou.org:5555;
-            index                   index.html              index.htm;
-        }
-
-        location  /webjars  {
-       	    proxy_pass              https://api.laokou.org:5555;
-            index                   index.html              index.htm;
-        }
-
-        location  /swagger-resources  {
-            proxy_pass              https://api.laokou.org:5555;
-            index                   index.html              index.htm;
-        }
-
-        location  /v3  {
-            proxy_pass              https://api.laokou.org:5555;
-            index                   index.html              index.htm;
-        }
-
         location  /  {
             # 你的打包路径
             root                    /usr/share/nginx/dist;
@@ -80,14 +60,14 @@ http  {
         }
 
         location @405 {
-            proxy_pass              https://api.laokou.org:5555$request_uri;
+            proxy_pass              https://gateway:5555$request_uri;
             proxy_set_header        Host                    $host;
             proxy_set_header        X-Real-IP               $remote_addr;
             proxy_set_header        X-Forwarded-For         $proxy_add_x_forwarded_for;
         }
 
-		location  /laokou/  {
-            proxy_pass              https://api.laokou.org:5555/;
+		location  /api/  {
+            proxy_pass              https://gateway:5555/;
             proxy_redirect          http://                 https://;
             proxy_connect_timeout   6000;
             proxy_read_timeout      6000;
