+++ ui/src/services/auth/auth.ts
@@ -29,3 +29,18 @@ export async function login(
 ) {
 	return token(params, options)
 }
+
+/** 清除令牌 清除令牌 DELETE /v3/logouts */
+export async function logoutV3(
+	body: API.LogoutParam,
+	options?: { [key: string]: any },
+) {
+	return request<any>('/api/auth/v3/tokens', {
+		method: 'DELETE',
+		headers: {
+			'Content-Type': 'application/json',
+		},
+		data: body,
+		...(options || {}),
+	});
+}
