+++ ui/src/services/auth/captcha.ts
@@ -19,12 +19,14 @@ export async function getCaptchaImageByUuidV3(
 /** 根据UUID发送验证码 根据UUID发送验证码 POST /v3/captchas */
 export async function sendCaptchaV3(
 	body: API.SendCaptchaParam,
+	requestId: string,
 	options?: { [key: string]: any },
 ) {
 	return request<any>('/api/auth/v3/captchas', {
 		method: 'POST',
 		headers: {
-			'Content-Type': 'application/json',
+			'request-id': requestId,
+			'Content-Type': 'application/json'
 		},
 		data: body,
 		...(options || {}),
